// api/app/AppServiceObjs.ts
export interface WeeklyReportCard {
  
  date: number
  date_from:number
  date_to:number
  is_read: boolean
  report_url: string
}

export interface RetrieveUserReportsResp {

  daily_report: DailyReportCard[]
  weekly_report: WeeklyReportCard[]
  num_of_new_weekly_report: number
}

export interface RetrieveOrCreateUserReportReq {

  date: number
}
// ...
export interface RetrieveFoodDiaryReq {

  date: number
}